Version "letfoxeswin"
Changes in find_dofile() concerning doing of dofiles. Layered
approach - * is done by *.do, *.do is done by *.do.do and so
on. Filenames are stripped until they will consist of "default"
prefix and any number N of ".do" suffices, but no further. Dofile
including N suffices can be done only by another dofile with
N + 1 suffices. If the search in some directory fails, it is
continued in the up-dir, an so forth until root dir.